abstact menu category
