+++ app/src/Menu/index.tsx
@@ -1,0 +1,47 @@
+/* @license Copyright 2023 @paritytech/polkadot-cloud authors & contributors
+SPDX-License-Identifier: GPL-3.0-only */
+
+import { Fragment } from "react";
+import { Separator } from "@packages/cloud-react/lib/base/structure/Separator";
+import { Link, useLocation } from "react-router-dom";
+import { nameFromRoute, routeCategories } from "../config/routes";
+import "../styles/menu.scss";
+import { Category } from "./Category";
+
+export const Menu = () => {
+  const { pathname } = useLocation();
+
+  return (
+    <div className="menu">
+      {routeCategories.map(({ name, ...rest }, i) => (
+        <Fragment key={`nav_${i}`}>
+          {"paths" in rest ? (
+            <>
+              <h3
+                style={{
+                  fontWeight: "bold",
+                }}
+              >
+                {name}
+              </h3>
+
+              <Category rest={rest} i={i} />
+            </>
+          ) : (
+            <>
+              <Link
+                className={`link lg${
+                  pathname === `/${rest.path}` ? ` selected` : ``
+                }`}
+                to={`/${rest.path}`}
+              >
+                {nameFromRoute(rest.path)}
+              </Link>
+            </>
+          )}
+          <Separator />
+        </Fragment>
+      ))}
+    </div>
+  );
+};
