+++ app/src/styles/menu.scss
@@ -6,7 +6,7 @@ SPDX-License-Identifier: GPL-3.0-only */
   top: 0;
   left: 0;
   height: 100vh;
-  padding-top: 5.5rem;
+  padding-top: 6rem;
   padding-left: 2rem;
   max-height: 100vh;
   width: 10rem;
@@ -31,10 +31,13 @@ SPDX-License-Identifier: GPL-3.0-only */
   }
 
   > h3 {
-    margin-bottom: 0;
+    margin-bottom: 0.2rem;
   }
 
   section {
+    transition: height 0.2s ease-out;
+    overflow: hidden;
+
     h5 {
       margin-top: 1.5rem;
     }
